fix tests with callback
