+++ .eslintrc.js
@@ -89,7 +89,7 @@ module.exports = {
         'max-classes-per-file': 'error',
         'max-depth': 'error',
         'max-len': 'off',
-        'max-lines': 'error',
+        'max-lines': 'off',
         'max-lines-per-function': 'off',
         'max-nested-callbacks': 'error',
         'max-params': 'error',
