+++ package.json
@@ -12,7 +12,7 @@
     "gen-doc": "./node_modules/.bin/apidoc -i src/ -o doc/",
     "open-doc": "node ./scripts/open.js",
     "doc": "npm run gen-doc && npm run open-doc",
-    "lint": "./node_modules/.bin/eslint \"src/**\" \"tests/**\" --fix"
+    "lint": "./node_modules/.bin/eslint src/** tests/** --fix"
   },
   "keywords": [
     "Xendit"
