+++ tests/clear-table.js
@@ -1,11 +1,14 @@
 'use strict';
 
-module.exports = (db) => {
+module.exports = (db, done) => {
     const clearTable = 'DELETE FROM Rides';
     const resetSequence = 'UPDATE sqlite_sequence SET seq = 0 WHERE name="Rides"';
 
-    db.run(clearTable);
-    db.run(resetSequence);
+    db.run(clearTable, () => {
+        db.run(resetSequence, () => {
+            done();
+        });
+    });
 
     return db;
 };
